refactor(Masonry): [Workspace] Move Workspace logic to Workspace View
